Default Xtalk scan now 2x1 or 2x2, same structure as ToT

+++ src/libFe65p2/include/Fe65p2MaskLoop.h
@@ -13,10 +13,10 @@
 class Fe65p2MaskLoop : public LoopActionBase {
     public:
         Fe65p2MaskLoop();
-        void setMaskStage(uint32_t mask);
+        void setMaskStage(uint16_t mask);
 
     private:
-        uint32_t m_mask;
+        uint16_t m_mask;
         unsigned m_cur;
 
         void init();
